@@ -67,7 +67,7 @@
         <junit.version>4.12</junit.version>
 
         <!-- java版本和文件编码 -->
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <file.encoding>UTF-8</file.encoding>
 
         <!-- maven核心插件 -->
@@ -185,41 +185,6 @@
                     <forkMode>once</forkMode>
                     <threadCount>10</threadCount>
                     <argLine>-Dfile.encoding=${file.encoding}</argLine>
-                </configuration>
-            </plugin>
-
-            <!-- 打包插件-->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>${maven-shade-plugin.version}</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <createDependencyReducedPom>false</createDependencyReducedPom>
-                        </configuration>
-                    </execution>
-                </executions>
-                <configuration>
-
-                    <transformers>
-                        <transformer
-                                implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                            <!--<mainClass>com.xxg.Main</mainClass>-->
-                        </transformer>
-                        <transformer
-                                implementation="org.apache.maven.plugins.shade.resource.AppendingTransformer">
-                            <resource>META-INF/spring.handlers</resource>
-                        </transformer>
-                        <transformer
-                                implementation="org.apache.maven.plugins.shade.resource.AppendingTransformer">
-                            <resource>META-INF/spring.schemas</resource>
-                        </transformer>
-                    </transformers>
                 </configuration>
             </plugin>
 
