@@ -68,7 +68,7 @@
         <junit.version>4.12</junit.version>
 
         <!-- java版本和文件编码 -->
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <file.encoding>UTF-8</file.encoding>
 
         <!-- maven核心插件 -->
